servers: Move server starters to their own .rs file to fix clutter in lib.rs.
