+++ core/tests/flow.rs
@@ -6,15 +6,16 @@ use bitcoin::{Address, Amount};
 use bitcoincore_rpc::Auth;
 use clementine_circuits::constants::BRIDGE_AMOUNT_SATS;
 use clementine_core::actor::Actor;
+use clementine_core::create_test_database;
 use clementine_core::db::common::Database;
 use clementine_core::extended_rpc::ExtendedRpc;
 use clementine_core::mock::common;
 use clementine_core::script_builder::ScriptBuilder;
+use clementine_core::servers::*;
 use clementine_core::traits::rpc::OperatorRpcClient;
 use clementine_core::transaction_builder::{CreateTxOutputs, TransactionBuilder};
 use clementine_core::utils::handle_taproot_witness_new;
 use clementine_core::EVMAddress;
-use clementine_core::{create_test_database, start_operator_and_verifiers};
 use std::thread;
 
 #[tokio::test]
